native/requirements.txt
PyQt6
PyQt6-WebEngine
requests
psutil
pywin32
Pillow

